Much simpler code coverage, similar to AFL
There's a global map, indexed by branch hashes, where we record if a branch has been seen or not. If a mutated input triggers a new branch, add it to the corpus.
Now the number displayed in coverage is the number of triggered branches, and it is different from the corpus number of elements.