Smaller coverage map, JIT perf improvements
Whenever JIT detects a fault, it always goes through the same fault path and same fast vm exit. It doesn't provide any extra information about the fault, so we run the interpreter to get it
With this perf improvement, now JIT runs at ~750 minstr/s, which is ~5.7 faster than interpreter!